refactor: change naming style to kebab
